# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brokers:
    get:
      operationId: listBrokers
      tags: [brokers]
      summary: List all broker connections
      responses:
        "200":
          description: List of broker connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrokerConnection"
    post:
      operationId: createBroker
      tags: [brokers]
      summary: Add a broker connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrokerBody"
      responses:
        "201":
          description: Broker connection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerConnection"
  /brokers/{id}:
    get:
      operationId: getBroker
      tags: [brokers]
      summary: Get a broker connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Broker connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerConnection"
        "404":
          description: Not found
    patch:
      operationId: updateBroker
      tags: [brokers]
      summary: Update broker credentials or settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrokerBody"
      responses:
        "200":
          description: Updated broker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerConnection"
    delete:
      operationId: deleteBroker
      tags: [brokers]
      summary: Remove a broker connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /brokers/{id}/test:
    post:
      operationId: testBrokerConnection
      tags: [brokers]
      summary: Test broker API connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"

  # ── Trading Strategies ────────────────────────────────────────────
...
  /bot/start:
    post:
      operationId: startBot
      tags: [bot]
      summary: Start the trading bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartBotBody"
      responses:
        "200":
          description: Bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
...
  /bot/logs:
    get:
      operationId: getBotLogs
      tags: [bot]
      summary: Get recent bot activity logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Bot logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BotLog"

  # ── Positions ─────────────────────────────────────────────────────
...
  /reports/performance:
    get:
      operationId: getPerformanceReport
      tags: [reports]
      summary: Get performance report for a period
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly, quarterly, annually]
        - name: brokerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: strategyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Performance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceReport"
  /reports/pnl-chart:
    get:
      operationId: getPnlChart
      tags: [reports]
      summary: Get P&L chart data points
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly, quarterly, annually]
        - name: brokerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: P&L time series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PnlDataPoint"
  /reports/win-rate:
    get:
      operationId: getWinRateStats
      tags: [reports]
      summary: Get win/loss statistics
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly, quarterly, annually]
      responses:
        "200":
          description: Win rate stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WinRateStats"
  /reports/top-symbols:
    get:
      operationId: getTopSymbols
      tags: [reports]
      summary: Get best and worst performing symbols
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly, quarterly, annually]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top performing symbols
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SymbolPerformance"

  # ── Dashboard ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    BrokerConnection:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brokerType:
          type: string
          description: "schwab | robinhood | alpaca | interactive_brokers"
        status:
          type: string
          description: "connected | disconnected | error"
        accountId:
          type: ["string", "null"]
        accountValue:
          type: ["number", "null"]
        buyingPower:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, brokerType, status, isActive, createdAt, updatedAt]
    CreateBrokerBody:
      type: object
      properties:
        name:
          type: string
        brokerType:
          type: string
        apiKey:
          type: ["string", "null"]
        apiSecret:
          type: ["string", "null"]
        accessToken:
          type: ["string", "null"]
        refreshToken:
          type: ["string", "null"]
        accountId:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [name, brokerType]
    UpdateBrokerBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        apiKey:
          type: ["string", "null"]
        apiSecret:
          type: ["string", "null"]
        accessToken:
          type: ["string", "null"]
        refreshToken:
          type: ["string", "null"]
        accountId:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    ConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        accountId:
          type: ["string", "null"]
        accountValue:
          type: ["number", "null"]
      required: [success, message]

    # ── Strategy ────────────────────────────────────────────────────
...
    BotStatus:
      type: object
      properties:
        isRunning:
          type: boolean
        startedAt:
          type: ["string", "null"]
        activeStrategyId:
          type: ["integer", "null"]
        activeBrokerId:
          type: ["integer", "null"]
        tradesExecutedToday:
          type: integer
        dailyPnl:
          type: number
        lastHeartbeat:
          type: ["string", "null"]
      required: [isRunning, tradesExecutedToday, dailyPnl]
    StartBotBody:
      type: object
      properties:
        strategyId:
          type: integer
        brokerId:
          type: integer
      required: [strategyId, brokerId]
    BotLog:
      type: object
      properties:
        id:
          type: integer
        level:
          type: string
          description: "info | warn | error | trade"
        message:
          type: string
        symbol:
          type: ["string", "null"]
        action:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, level, message, createdAt]

    # ── Positions ───────────────────────────────────────────────────
...
    PerformanceReport:
      type: object
      properties:
        period:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        totalTrades:
          type: integer
        winningTrades:
          type: integer
        losingTrades:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        totalPnlPercent:
          type: number
        avgWin:
          type: number
        avgLoss:
          type: number
        profitFactor:
          type: number
        maxDrawdown:
          type: number
        sharpeRatio:
          type: ["number", "null"]
        bestTrade:
          type: ["number", "null"]
        worstTrade:
          type: ["number", "null"]
      required: [period, startDate, endDate, totalTrades, winningTrades, losingTrades, winRate, totalPnl, totalPnlPercent, avgWin, avgLoss, profitFactor, maxDrawdown]
    PnlDataPoint:
      type: object
      properties:
        date:
          type: string
        pnl:
          type: number
        cumulativePnl:
          type: number
        tradeCount:
          type: integer
      required: [date, pnl, cumulativePnl, tradeCount]
    WinRateStats:
      type: object
      properties:
        winRate:
          type: number
        wins:
          type: integer
        losses:
          type: integer
        breakeven:
          type: integer
        avgWinAmount:
          type: number
        avgLossAmount:
          type: number
        largestWin:
          type: number
        largestLoss:
          type: number
      required: [winRate, wins, losses, breakeven, avgWinAmount, avgLossAmount, largestWin, largestLoss]
    SymbolPerformance:
      type: object
      properties:
        symbol:
          type: string
        totalPnl:
          type: number
        totalPnlPercent:
          type: number
        tradeCount:
          type: integer
        winRate:
          type: number
      required: [symbol, totalPnl, totalPnlPercent, tradeCount, winRate]

    # ── Dashboard ───────────────────────────────────────────────────
...